Clean up `LinearMap`, address pull request comments
Remove `RandomAccessIterator` implementations, because the underlying type
isn't logically random-access. Keep `DoubleEndedIterator` implementation
because it is needed for `ExactSizeIterator`.

Flatten iterator implementations for proxy structs.

Add remaining docs, warn on missing docs.

Remove the comment about the missing `Copy` implementation, the status on this
seems to be that `Copy` is not implemented for types where this might lead to
usability issues.