Buffer: a safe utility collection for allocating memory.
Buffer provides an entirely safe interface for managing memory within other
collections. It provides allocate, reallocate, and deallocate functionality
and obeys RAII principles (deallocate on drop).

This can be used to simplify the implementation of any data structure which
does allocation not through another structure like Vec.

From the docs:

A safe wrapper around a heap allocated buffer of Ts, tracking capacity only.

Buffer makes no promises about the actual contents of this memory, that's up
to the user of the structure and can be manipulated using the standard pointer
utilities, accessible through the impl of `Deref<Target=*mut T>` for `Buffer<T>`.

As a result of this hands-off approach, `Buffer`s destructor does not attempt
to drop any of the contained elements; the destructor simply frees the contained
memory.

You can think of `Buffer<T>` as an approximation for `Box<[T]>` where the elements
are not guaranteed to be valid/initialized. It is meant to be used as a building
block for other collections, so they do not have to concern themselves with the
minutiae of allocating, reallocating, and deallocating memory.