Buffer: Fix an incorrect free caused by capacity overflow during reallocation.
By only swapping out the buffer and not also changing the capacity
to 0 during reallocation, a capacity overflow could cause the drop
impl of Buffer to run when the buffer was empty() and the capacity
was non-zero, causing an incorrect call to deallocate.

We now also swap the capacity to 0, so that when reallocate is called
(the site of a possible panic) the buffer is in a consistent empty
state.

Also, in order to avoid leaking memory on oom (ya kind of useless,
but still) reallocate now deallocates the old ptr if reallocation
fails.