correct comparators' `PhantomData` fields
Comparators are essentially functions that accept two immutable
references, so the use of `PhantomData<*mut T>` was overly conservative.
They never have access to a mutable `T`, so the `PhantomData<fn(&T)>`
formulation models the situation most accurately.

This has the side effect of making `Natural` and `Borrow` implement
`Send` and `Sync`, which is what one would expect.